Fix a mistake in aircraft plotting
The idea of the plot_aircraft() method is that if an aircraft flies
directly overhead the receiver, the pixel will be very bright but
shaded the same as the aircraft.

This didn't work as the rcvr variable was set to the GPS coordinates
of the receiver instead of pixel coordinates, and the conditional never
matched.