Removing 'retart' subparser to correctly disable 'restart' action
Currently `restart` action is not working, as each time `radarscoped` is
executed, it spawns a new proces which doesn't currently have any means
to communicate with the actual running daemon.

The `restart` action, as well as some of the more advanced features of
`status` will need to be re-designed. Disabling for now.